backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
python-multipart==0.0.6
qdrant-client>=1.9.0

# Document processing
pymupdf
Pillow
aiofiles

# ColQwen2 / ColPali
colpali-engine
torch
torchvision
transformers

# Qwen2.5-VL vision LLM
qwen-vl-utils
bitsandbytes
